move default genesis to genesis package
